fix(views): stamp GPU cells with the same labels and gnomon as CPU cells
The GPU path composited raw PNGs. No view name, no axis gnomon. So a sheet
produced by the render port silently dropped the model's orientation cues,
and which one it got depended on whether a GPU happened to be reachable.

I flagged this as arguably a useful degrade tell and left it. That was the
wrong call: producer identity is already reported as `viewsRendererId` and
`renderDegraded`, which a consumer can actually read. Encoding it as "the
picture has no words on it" costs the model context on exactly the path where
nothing tells it that happened.

Both producers now call annotateViewCell. It lives in its own leaf module
(with stampAxisGnomon moved out of views/index.ts) so the grid compositor can
reach it without importing the rasterizer, and so the label scale, corner
anchors and gnomon basis are written once rather than twice. A host returns
pixels and knows nothing of the Kiln camera vocabulary, so the annotation has
to happen engine-side or not at all.

compositeViewPngGrid takes an optional views[] and throws if it does not align
1:1 with the PNGs — a misaligned list would label cells with the wrong camera,
which is worse than not labelling them.

Custom cell names reach the GPU sheet too, so a model that names a cell "seam"
sees "SEAM" on it regardless of which renderer ran.

1025 pass / 0 fail, typecheck and format clean, no new lint findings.
